requirements.txt
fastapi~=0.95.2
uvicorn~=0.22.0
random_user_agent~=1.0.1
requests~=2.31.0
bs4~=0.0.1
beautifulsoup4~=4.12.2
pydantic~=1.10.8
starlette~=0.27.0
torch~=2.0.1
torchvision~=0.15.2
torchaudio~=2.0.2
httpx~=0.24.1
pandas~=2.0.2
tqdm~=4.65.0
pyarrow~=12.0.0
numpy~=1.24.3
transformers~=4.30.0
Pillow~=9.5.0